+++ crawler/src/RequestManager.cpp
@@ -94,7 +94,7 @@ void RequestManager::Run(ThreadSync& sync) {
         auto& ready = requestExecutor_.ReadyResponses();
         if (!ready.empty()) {
             for (const auto& r : ready) {
-                limiter_->UnleaseHost(r.req.Url().host);
+                limiter_->UnleaseHost(r.req.Url().host, r.req.Url().NonEmptyPort());
             }
             docQueue_->PushAll(ready);
             ready.clear();
@@ -118,7 +118,7 @@ void RequestManager::TouchRequestTimeouts() {
 }
 
 void RequestManager::DispatchFailedRequest(const http::FailedRequest& failed) {
-    limiter_->UnleaseHost(failed.req.Url().host);
+    limiter_->UnleaseHost(failed.req.Url().host, failed.req.Url().NonEmptyPort());
 
     auto s = std::string{http::StringOfRequestError(failed.error)};
     spdlog::warn("failed crawl request: {} {}", failed.req.Url().url, s);